feat(social): complete flow for inviting and agreeing coop friend, except disagreement.
